LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-image-href.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><image href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=b69c804a-0c45-4758-bea8-fa8dde67aa86&amp;encodingcheck=&Gbreve;"></image></svg>


FAIL Speculative parsing, document.write(): svg-image-href Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
